Rename db module to database
