+++ backend/src/database/models/Artist.model.ts
