+++ backend/src/database/models/ArtistGenre.model.ts
