+++ backend/src/schema/resolvers/GenreResolver.ts
@@ -1,6 +1,6 @@
 import { Op, Sequelize } from 'sequelize'
-import ArtistGenre from '../../db/models/ArtistGenre.model'
-import Artist from '../../db/models/Artist.model'
+import ArtistGenre from '../../database/models/ArtistGenre.model'
+import Artist from '../../database/models/Artist.model'
 import {
   IFindGenresLikeArgs,
   IGetArtistsFromGenreArgs
