+++ backend/src/schema/types/SongType.ts
@@ -4,7 +4,7 @@ export default gql`
   type Song {
     id: String!
     artistId: String!
-    artist: Artist!
+    artists: [Artist!]
     name: String!
     album: String!
     art: String
